Move client to MLV folder
