+++ .gitignore
@@ -5,3 +5,4 @@ todo.txt
 .coverage
 htmlcov
 run_tests.bat
+local
